Factor out match logic into abstract _match method
